+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sl-design-system/angular
 
+## 1.0.1
+
+### Patch Changes
+
+- [#1296](https://github.com/sl-design-system/components/pull/1296) [`39a4cb2`](https://github.com/sl-design-system/components/commit/39a4cb206ad923862c902b3ac7dddd4ae5b87746) - Update to use the new TextField generic type
+
 ## 1.0.0
 
 ### Major Changes